fix settings routing
