+++ lib/settings-routes/index.js
@@ -1,0 +1,13 @@
+import page from 'page'
+
+/**
+ * Reload settings page
+ */
+
+page('/settings', reload)
+page('/settings/', reload)
+page('/settings/*', reload)
+
+function reload () {
+  window.location.reload(false)
+}
