+++ lib/site/boot/boot.js
@@ -20,22 +20,10 @@ import '../../debug'
 let log = debug('democraycos:boot')
 
 /**
- * Reload settings page
- */
-
-page('/settings', reload)
-page('/settings/', reload)
-page('/settings/*', reload)
-
-function reload () {
-  debugger
-  window.location.reload(false)
-}
-
-/**
  * Mount applications.
  */
 
+import '../../settings-routes'
 import '../../analytics'
 import '../../body-classes/body'
 import '../../content-lock/locker'
